:sparkles: feature: 모집 게시글 필터 적용
무한 스크롤 스크롤이 화면 상단으로 튕기는 현상이 있어, ArticleList 컴포넌트를 빼지 않고 페이지 내에 위치하도록 수정
